// src/app/models/unayoe/mi-perfil-unayoe/miPerfilUnayoe.model.ts
export interface DatosMiPerfilUnayoe {
    id: number;
    nombre: string;
    apellido_paterno: string;
    apellido_materno: string;
    sexo: number;
    profesion: string;
    facebook: string;
    celular: string;
    correoAlternativo: string;
    wsp: string;
    foto: string;
    auto_descripcion: string;
    facultad: string;
    correoPrincipal: string;
    estado: number;
    rol: string
}