// src/app/models/unayoe/mi-perfil-unayoe/datosActualizarUnayoe.model.ts
export interface DatosActualizarUnayoe {
    id: number;
    nombre: string;
    apellido_paterno: string;
    apellido_materno: string;
    sexo: number;
    profesion: string;
    facebook: string;
    celular: string;
    correoAlternativo: string;
    wsp: string;
    foto: string;
    auto_descripcion: string;
}